requirements.txt
psycopg2-binary
cryptography
python-dotenv
fastapi
uvicorn
transformers
torch
valkey

opentelemetry-api
opentelemetry-sdk
opentelemetry-instrumentation-fastapi
opentelemetry-instrumentation-psycopg2
opentelemetry-instrumentation-requests
opentelemetry-exporter-otlp
